requirements.txt
Pillow
streamlit
streamlit_option_menu